+++ docs/ChangeLog.md
@@ -127,6 +127,7 @@
 - syntax: highlight `\?` in here documents `#D1959` e619e73
 - syntax: recognize history expansion in here documents, `"...!"` (bash <= 4.2), and `$!` (bash <= 4.1) `#D1959` e619e73
 - syntax: support context after `((...))` and `[[ ... ]]` in bash-5.2 `#D1962` 67cb967
+- edit: support the readline variable `search-ignore-case` of bash-5.3 `#D1976` xxxxxxx
 
 ## Changes